requirements.txt
fastapi
uvicorn
requests
pandas
numpy
yfinance
beautifulsoup4
pyttsx3
transformers
sentence-transformers
faiss-cpu
streamlit
whisper
sounddevice
wavio
